+++ code/engine.vc2008/xrGame/ai/monsters/basemonster/base_monster.h
@@ -472,7 +472,7 @@ public:
 #endif
 
 #ifdef DEBUG
-	virtual void			debug_on_key		(int key) {}
+	virtual void			debug_on_key		(u8 key) {}
 #endif
 //////////////////////////////////////////////////////////////////////////
